requirements_pip.txt
pytest
pytest-cov
python-dateutil
Flask
Werkzeug
gunicorn
connexion[swagger-ui]
marshmallow
mongoengine
marshmallow-mongoengine
python-dateutil
python-qpid_proton
git+https://git@github.com/eurocontrol-swim/rest-client.git
git+https://git@github.com/eurocontrol-swim/swim-backend.git
git+https://git@github.com/eurocontrol-swim/swim-qpid-proton.git
git+https://git@github.com/eurocontrol-swim/pubsub-facades.git
git+https://git@github.com/eurocontrol-swim/subscription-manager-client.git
